AssignCodewords fix double overflow of marker[1]
If you do not have the break where I put it, `marker[1]` would get executed twice. So from `marker[1] == 1`, it would first become `marker[1] == 2` (10_b) and then `marker[1] == 3` (11_b), which is probably not intended. This might not matter, as it only happens once the leaf is overful, i.e. fully specified, but it would allow for a simpler underspecify check at the end (which currently is missing).